pyproject.toml
[project]
name = "TLE"
version = "0.1.0"
requires-python = ">= 3.10"
dependencies = [
  # Discord bot core
  "discord.py >= 2.3, < 3",
  "aiocache >= 0.12, < 1",
  # Data & plotting
  "numpy >= 1.24, < 3",
  "pandas >= 2.0, < 3",
  "matplotlib >= 3.7, < 4",
  "seaborn >= 0.12, < 1",
  # Images
  "pycairo >= 1.24, < 2",
  "PyGObject >= 3.42, < 4",
  # Database
  "aiosqlite >= 0.19, < 1",
  # Config
  "python-dotenv >= 1.0, < 2",
  # OAuth / JWT
  "PyJWT >= 2.8, < 3",
]

[project.optional-dependencies]
test = [
  "pytest >= 7.0, < 9",
  "pytest-asyncio >= 0.23, < 1",
  "pytest-cov >= 4.0, < 6",
  "pytest-mock >= 3.10, < 4",
]

[tool.setuptools.packages.find]
include = ["tle*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true
follow_imports = "skip"
show_error_codes = true
show_column_numbers = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring mocked Discord context",
]
